Fixed numerous errors regarding improperly serialized boards.
Added Boolean fields to the King class to denote if he can castle or
not, as well as additional utility methods to determine castling
eligibility.

Added a toString() method in the board to help with serialization, as
well as a field to contain the en passant tile, even if there are no
pawns that can use it.

Made a slightly better AI that may run slower. It not only takes into
account material, but position. Values are calculated for both players
instead of just one.